Update gabbits to reflect progress in nested providers
We now have nested providers in GET /allocation_candidates
so the provider summaries provided are more detailed and
results are more complete. The changes here update the
comments and the tests to show off some of those features.

Bugs remain with writing allocations for consumers that were
created, has their allocaitons deleted, and then want more
allocations written. This is being fixed, refrences are
left in the files.