Add indicator of build failure
And intentionally make things fail. The idea here is wanting
a visible track of the state of things. If placecat isn't
working right, then placement is not currently filling its
promises.

To help with that, a trigger has been set up so that when
placedock has a new build, that triggers a new build here
on placecat.